[IMP] im_livechat: remove ThreadTypingMixin (merge with PublicLivechat)
Task-2894695
